+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Automate zipping plugin with gulp
 - Adapt styles to wp-scripts update
 - Specify package and text domain
+- Fix ListView name substitution on production build
 
 ## [v0.1.0-alpha.2] - 2020-08-07
 - improved WP FullCalendar migration